refactor: move key pair code
