+++ packages/core/src/key_pair/constants.ts
@@ -1,0 +1,4 @@
+/** All supported key types */
+export enum KeyType {
+    ED25519 = 0,
+}
