+++ packages/core/src/key_pair/index.ts
@@ -1,0 +1,4 @@
+export { KeyType } from './constants';
+export { KeyPair, Signature } from './key_pair';
+export { KeyPairEd25519 } from './key_pair_ed25519';
+export { PublicKey } from './public_key';
