+++ packages/core/src/key_pair/key_pair_ed25519.ts
@@ -1,0 +1,59 @@
+import { baseEncode, baseDecode } from 'borsh';
+import nacl from 'tweetnacl';
+
+import { KeyType } from './constants';
+import { KeyPair, Signature } from './key_pair';
+import { PublicKey } from './public_key';
+
+/**
+ * This class provides key pair functionality for Ed25519 curve:
+ * generating key pairs, encoding key pairs, signing and verifying.
+ */
+export class KeyPairEd25519 extends KeyPair {
+    readonly publicKey: PublicKey;
+    readonly secretKey: string;
+
+    /**
+     * Construct an instance of key pair given a secret key.
+     * It's generally assumed that these are encoded in base58.
+     * @param {string} secretKey
+     */
+    constructor(secretKey: string) {
+        super();
+        const keyPair = nacl.sign.keyPair.fromSecretKey(baseDecode(secretKey));
+        this.publicKey = new PublicKey({ keyType: KeyType.ED25519, data: keyPair.publicKey });
+        this.secretKey = secretKey;
+    }
+
+    /**
+     * Generate a new random keypair.
+     * @example
+     * const keyRandom = KeyPair.fromRandom();
+     * keyRandom.publicKey
+     * // returns [PUBLIC_KEY]
+     *
+     * keyRandom.secretKey
+     * // returns [SECRET_KEY]
+     */
+    static fromRandom() {
+        const newKeyPair = nacl.sign.keyPair();
+        return new KeyPairEd25519(baseEncode(newKeyPair.secretKey));
+    }
+
+    sign(message: Uint8Array): Signature {
+        const signature = nacl.sign.detached(message, baseDecode(this.secretKey));
+        return { signature, publicKey: this.publicKey };
+    }
+
+    verify(message: Uint8Array, signature: Uint8Array): boolean {
+        return this.publicKey.verify(message, signature);
+    }
+
+    toString(): string {
+        return `ed25519:${this.secretKey}`;
+    }
+
+    getPublicKey(): PublicKey {
+        return this.publicKey;
+    }
+}
